refactor: move CodePoint and Encoding enums to separate files within same namespace
