+++ src/Danaus/Url/Encoding.cs
@@ -1,0 +1,6 @@
+namespace Danaus.Url;
+
+public enum Encoding
+{
+    UTF8
+}
