+++ src/Danaus/Url/URLParser.cs
@@ -3,11 +3,6 @@ using System.Text;
 
 namespace Danaus.Url;
 
-public enum Encoding
-{
-    UTF8
-}
-
 public enum PercentEncodeSet
 {
     C0Control,
@@ -43,58 +38,6 @@ public enum State
     SpecialAuthorityIgnoreSlashes,
     SpecialAuthoritySlashes,
     SpecialRelativeOrAuthority,
-}
-
-public enum CodePoint
-{
-    Ampersand = 0x0026,
-    Asterisk = 0x002A,
-    Apostrophe = 0x0027,
-    ApplicationProgramCommand = 0x009F,
-    CarriageReturn = 0x000D,
-    CircumflexAccent = 0x005E,
-    Colon = 0x003A,
-    Comma = 0x002C,
-    CommercialAt = 0x0040,
-    Delete = 0x007F,
-    DollarSign = 0x0024,
-    EqualsSign = 0x003D,
-    ExclamationMark = 0x0021,
-    GraveAccent = 0x0060,
-    GreaterThanSign = 0x003E,
-    HyphenMinus = 0x002D,
-    LeftCurlyBracket = 0x007B,
-    LeftParenthesis = 0x0028,
-    LeftSquareBracket = 0x005B,
-    LessThanSign = 0x003C,
-    LineFeed = 0x000A,
-    LowercaseA = 0x0061,
-    LowercaseF = 0x0066,
-    LowercaseZ = 0x007A,
-    LowLine = 0x005F,
-    Nine = 0x0039,
-    NullCharacter = 0x0000,
-    Number = 0x0023,
-    PercentSign = 0x0025,
-    Period = 0x002E,
-    Plus = 0x002B,
-    QuestionMark = 0x003F,
-    QuotationMark = 0x022,
-    ReverseSolidus = 0x005C,
-    RightCurlyBracket = 0x007D,
-    RightParenthesis = 0x0029,
-    RightSquareBracket = 0x005D,
-    Semicolon = 0x003B,
-    Solidus = 0x002F,
-    Space = 0x0020,
-    Tab = 0x0009,
-    Tilde = 0x007E,
-    UnitSeparator = 0x001F,
-    UppercaseA = 0x0041,
-    UppercaseF = 0x0046,
-    UppercaseZ = 0x005A,
-    VerticalLine = 0x007C,
-    Zero = 0x0030,
 }
 
 public class URLParser
